extract SplashRouter code to production
